+++ lib/install.ps1
@@ -1211,3 +1211,17 @@ function persist_permission($manifest, $global) {
         $acl | Set-Acl -Path $path
     }
 }
+
+# test if there are running processes
+function test_running_process($app, $global) {
+    $processdir = appdir $app $global | Convert-Path
+    $running_processes = Get-Process | Where-Object { $_.Path -like "$processdir\*" }
+
+    if ($running_processes -and !(get_config 'ignore_running_processes')) {
+        error "Application `"$app`" is still running. Close all instances and try again."
+        return $true
+    } else {
+        warn "Application `"$app`" is still running. Scoop is configured to ignore this condition."
+        return $false
+    }
+}